Make nnz be int64 type for COO matrices (#386)
* making nnz have type int64 in coo routines

* testing

* remove comments

* testing performance improvements to random coo matrix generation

* comment init code

* testing

* testing

* testing

* testing

* fixing copyright

* testing

* testing

* fixing formatting

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* tesitng

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* testing

* remove comments and debug statements

* remove comments

* PR fixes

* fixing formatting

* PR fixes

* PR fixes


[ROCm/rocSPARSEcommit: f89d89257b391e5d20c013cbb0a34867853b5006]

+++ library/src/level2/rocsparse_csrsv_analysis.cpp
@@ -190,7 +190,7 @@ rocsparse_status rocsparse_trm_analysis(rocsparse_handle          handle,
 
 // Run analysis
 #define CSRSV_DIM 1024
-    dim3 csrsv_blocks((handle->wavefront_size * m - 1) / CSRSV_DIM + 1);
+    dim3 csrsv_blocks(((int64_t)handle->wavefront_size * m - 1) / CSRSV_DIM + 1);
     dim3 csrsv_threads(CSRSV_DIM);
 
     if(trans == rocsparse_operation_none)
